% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ot.R
\name{aggregate_barycentric_weights}
\alias{aggregate_barycentric_weights}
\title{Combine neighbour plans into barycentric weights}
\usage{
aggregate_barycentric_weights(plan_rows, neighbor_idx, neighbor_dists)
}
\description{
Combine neighbour plans into barycentric weights
}
\keyword{internal}
