% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_framework.R
\name{aligner_capabilities}
\alias{aligner_capabilities}
\title{Capabilities for an aligner}
\usage{
aligner_capabilities(algo)
}
\arguments{
\item{algo}{An aligner object created by new_aligner or an aligner constructor}
}
\value{
A list with elements group (character) and supports_multi (logical)
}
\description{
Capabilities for an aligner
}
\examples{
algo <- new_aligner("test", group = "O")
caps <- aligner_capabilities(algo)
}
