% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp_multiset.R
\name{build_anchor}
\alias{build_anchor}
\title{Build Anchor Representation}
\usage{
build_anchor(desc, bases, Mlist, anchor)
}
\arguments{
\item{desc}{List of descriptor matrices}

\item{bases}{List of spectral bases}

\item{Mlist}{List of rotation matrices}

\item{anchor}{Index of anchor graph or "mean"}
}
\value{
Anchor coefficient matrix
}
\description{
Constructs the anchor coefficient matrix either from a specific graph
or as the mean of all aligned representations.
}
\keyword{internal}
