% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw.R
\name{classical_ot_lp}
\alias{classical_ot_lp}
\title{Classical OT linear program}
\usage{
classical_ot_lp(cost, p, q)
}
\value{
A numeric matrix representing the optimal transport plan.
}
\description{
Classical OT linear program
}
\keyword{internal}
