% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{coef.fpgw}
\alias{coef.fpgw}
\title{Extract coefficients from FPGW}
\usage{
\method{coef}{fpgw}(object, type = c("transport", "parameters"), ...)
}
\arguments{
\item{object}{An fpgw object}

\item{type}{Type of coefficients to extract}

\item{...}{Additional arguments}
}
\value{
Coefficients (transport plans or parameters)
}
\description{
Extract coefficients from FPGW
}
