% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_align.R
\name{compute_block_indices}
\alias{compute_block_indices}
\title{Compute block indices for multiblock structure}
\usage{
compute_block_indices(pdata)
}
\arguments{
\item{pdata}{Preprocessed data list}
}
\value{
Matrix with start and end indices for each block
}
\description{
Compute block indices for multiblock structure
}
\keyword{internal}
