% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_align.R
\name{compute_cone_embeddings}
\alias{compute_cone_embeddings}
\title{Compute Spectral Embeddings for CONE-Align}
\usage{
compute_cone_embeddings(strata, ncomp, sigma, use_laplacian, knn)
}
\arguments{
\item{strata}{List of data domains}

\item{ncomp}{Number of embedding dimensions}

\item{sigma}{Diffusion parameter}

\item{use_laplacian}{Whether to use Laplacian normalization}

\item{knn}{Number of nearest neighbors (NULL for adaptive)}
}
\value{
List of embedding matrices
}
\description{
Compute Spectral Embeddings for CONE-Align
}
\keyword{internal}
