% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_edge_distances_cpp}
\alias{compute_edge_distances_cpp}
\title{Compute pairwise squared distances for connected nodes}
\usage{
compute_edge_distances_cpp(X, A)
}
\arguments{
\item{X}{Feature matrix (n x d)}

\item{A}{Adjacency matrix (sparse, n x n)}
}
\value{
Matrix of squared distances for edges
}
\description{
Compute pairwise squared distances for connected nodes
}
