% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_edge_gradient_cpp}
\alias{compute_edge_gradient_cpp}
\title{Compute edge consistency gradient}
\usage{
compute_edge_gradient_cpp(S, A1, A2, X1, X2)
}
\arguments{
\item{S}{Current transport plan (n1 x n2)}

\item{A1}{Adjacency matrix of network 1 (sparse)}

\item{A2}{Adjacency matrix of network 2 (sparse)}

\item{X1}{Features of network 1 (n1 x d)}

\item{X2}{Features of network 2 (n2 x d)}
}
\value{
Gradient matrix (n1 x n2)
}
\description{
Efficient computation of the gradient of edge consistency regularizer
}
