% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot_proximal.R
\name{compute_edge_gradient_r}
\alias{compute_edge_gradient_r}
\title{Compute Gradient of Edge Consistency Regularizer (R implementation)}
\usage{
compute_edge_gradient_r(S, networks)
}
\arguments{
\item{S}{Current transport plan}

\item{networks}{Network structures}
}
\value{
Gradient matrix
}
\description{
Compute Gradient of Edge Consistency Regularizer (R implementation)
}
\keyword{internal}
