% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp.R
\name{compute_grasp_basis}
\alias{compute_grasp_basis}
\title{Block A: Spectral Basis Construction - Corrected}
\usage{
compute_grasp_basis(strata, ncomp, use_laplacian = TRUE)
}
\arguments{
\item{strata}{List of data domains}

\item{ncomp}{Number of spectral components}

\item{use_laplacian}{Whether to use Laplacian normalization}
}
\value{
List of spectral bases for each domain
}
\description{
Block A: Spectral Basis Construction - Corrected
}
\keyword{internal}
