% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp.R
\name{compute_grasp_descriptors}
\alias{compute_grasp_descriptors}
\title{Block B: Multi-scale Descriptors - Corrected}
\usage{
compute_grasp_descriptors(bases, q_descriptors, sigma = 0.73)
}
\arguments{
\item{bases}{List of spectral bases}

\item{q_descriptors}{Number of descriptor functions}

\item{sigma}{Diffusion parameter}
}
\value{
List of descriptor matrices
}
\description{
Block B: Multi-scale Descriptors - Corrected
}
\keyword{internal}
