% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_parrot_cost_cpp}
\alias{compute_parrot_cost_cpp}
\title{Compute PARROT cost matrix}
\usage{
compute_parrot_cost_cpp(
  networks,
  rwr_features,
  anchor_info,
  alpha,
  sigma,
  gamma
)
}
\arguments{
\item{alpha}{Weight for attribute vs RWR cost}

\item{sigma}{RWR restart probability}

\item{gamma}{Cross-graph discount factor}

\item{X1}{Features of network 1}

\item{X2}{Features of network 2}

\item{R1}{RWR descriptors of network 1}

\item{R2}{RWR descriptors of network 2}

\item{W1}{Transition matrix of network 1}

\item{W2}{Transition matrix of network 2}
}
\value{
Position-aware cost matrix
}
\description{
Complete cost matrix computation including Sylvester equation
}
