% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{compute_parrot_embeddings}
\alias{compute_parrot_embeddings}
\title{Compute PARROT Embeddings}
\usage{
compute_parrot_embeddings(networks, transport_plan, ncomp)
}
\arguments{
\item{networks}{List of network structures}

\item{transport_plan}{Transport plan matrix}

\item{ncomp}{Number of components for embedding}
}
\value{
Combined embedding matrix
}
\description{
Compute PARROT Embeddings
}
\keyword{internal}
