% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot_proximal.R
\name{compute_parrot_objective}
\alias{compute_parrot_objective}
\title{Compute PARROT objective}
\usage{
compute_parrot_objective(
  S,
  C_rwr,
  networks,
  anchor_info,
  lambda_e,
  lambda_n,
  lambda_p
)
}
\arguments{
\item{S}{Transport plan matrix}

\item{C_rwr}{Position-aware cost matrix}

\item{networks}{List of network structures}

\item{anchor_info}{Anchor information}

\item{lambda_e}{Edge consistency weight}

\item{lambda_n}{Neighborhood consistency weight}

\item{lambda_p}{Anchor prior weight}
}
\value{
Objective function value
}
\description{
Computes the full objective:
\eqn{<C_rwr, S> + lambda_e * L_e(S) + lambda_n * L_n(S) + lambda_p * L_p(S)}.
}
\keyword{internal}
