% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{compute_parrot_rwr_r}
\alias{compute_parrot_rwr_r}
\title{Compute RWR Features for PARROT (R implementation)}
\usage{
compute_parrot_rwr_r(networks, anchor_info, sigma, max_iter, tol)
}
\arguments{
\item{networks}{List of network structures}

\item{anchor_info}{List with anchor information for both networks}

\item{sigma}{RWR restart probability (beta in paper)}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}
}
\value{
List of RWR descriptor matrices
}
\description{
Compute RWR Features for PARROT (R implementation)
}
\keyword{internal}
