% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ot.R
\name{compute_point_distance}
\alias{compute_point_distance}
\title{Compute distances between a point and all rows of a matrix}
\usage{
compute_point_distance(point, data, metric, data_norm = NULL)
}
\value{
A numeric vector of distances from the query point to each row of `data`.
}
\description{
Compute distances between a point and all rows of a matrix
}
\keyword{internal}
