% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{compute_squared_distances}
\alias{compute_squared_distances}
\title{Compute Squared Euclidean Distance Matrix (Dispatcher)}
\usage{
compute_squared_distances(X1, X2, use_cpp = FALSE)
}
\arguments{
\item{X1}{First matrix (n1 x d)}

\item{X2}{Second matrix (n2 x d)}

\item{use_cpp}{Whether to use C++ implementation}
}
\value{
Distance matrix (n1 x n2)
}
\description{
Compute Squared Euclidean Distance Matrix (Dispatcher)
}
\keyword{internal}
