% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_squared_distances_cpp}
\alias{compute_squared_distances_cpp}
\title{Compute squared Euclidean distance matrix}
\usage{
compute_squared_distances_cpp(X1, X2)
}
\arguments{
\item{X1}{First matrix (n1 x d)}

\item{X2}{Second matrix (n2 x d)}
}
\value{
Distance matrix (n1 x n2)
}
\description{
Efficient computation of pairwise squared distances
}
