% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{compute_squared_distances_r}
\alias{compute_squared_distances_r}
\title{Compute Squared Euclidean Distance Matrix (R implementation)}
\usage{
compute_squared_distances_r(X1, X2)
}
\arguments{
\item{X1}{First matrix (n1 x d)}

\item{X2}{Second matrix (n2 x d)}
}
\value{
Distance matrix (n1 x n2)
}
\description{
Compute Squared Euclidean Distance Matrix (R implementation)
}
\keyword{internal}
