% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_align_multiple.R
\name{cone_align_multiple.list}
\alias{cone_align_multiple.list}
\title{CONE-Align Multiple for List of Matrices}
\usage{
\method{cone_align_multiple}{list}(data, ...)
}
\arguments{
\item{data}{A list containing three or more matrices (nodes x features)}

\item{...}{Additional arguments passed to hyperdesign method}
}
\description{
CONE-Align Multiple for List of Matrices
}
