% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_multi.R
\name{cone_multi_aligner}
\alias{cone_multi_aligner}
\title{CONE-Align multiple aligner (native multi-view)}
\usage{
cone_multi_aligner()
}
\value{
An object of class `c("cone_multi_aligner", "aligner")` representing
  a CONE-Align multiple alignment algorithm descriptor.
}
\description{
CONE-Align multiple aligner (native multi-view)
}
\examples{
algo <- cone_multi_aligner()
aligner_capabilities(algo)
}
