% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{.auto_select_sigma_P}
\alias{.auto_select_sigma_P}
\title{Auto-select sigma_P using histogram-spread heuristic}
\usage{
.auto_select_sigma_P(X, W0, T, M, verbose = FALSE)
}
\arguments{
\item{X}{Preprocessed data matrix}

\item{W0}{Initial weight matrix}

\item{T}{Target similarity matrix}

\item{M}{Mask matrix}

\item{verbose}{Logical for progress reporting}
}
\value{
Numeric scalar giving the optimal sigma_P value
}
\description{
Auto-select sigma_P using histogram-spread heuristic
}
\keyword{internal}
