% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{.handle_formula_interface}
\alias{.handle_formula_interface}
\title{Handle formula interface for supervised embedding}
\usage{
.handle_formula_interface(
  formula,
  data,
  T,
  M,
  sigma_P,
  ncomp,
  alpha_p,
  alpha_schedule,
  maxit,
  tol,
  batch_size,
  use_cpp,
  verbose,
  lr,
  ...
)
}
\value{
A simembed object with formula and response components added
}
\description{
Handle formula interface for supervised embedding
}
\keyword{internal}
