% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolabel.R
\name{.select_diverse_representatives}
\alias{.select_diverse_representatives}
\title{Select diverse cluster representatives}
\usage{
.select_diverse_representatives(
  sim_matrix,
  cluster_assignments,
  valid_clusters,
  diversity_weight,
  min_clusters,
  max_clusters,
  use_advanced_diversity = TRUE,
  verbose
)
}
\arguments{
\item{sim_matrix}{Original similarity matrix}

\item{cluster_assignments}{Vector of cluster assignments}

\item{valid_clusters}{Vector of valid cluster IDs}

\item{diversity_weight}{Weight for diversity constraint}

\item{min_clusters}{Minimum number of clusters}

\item{max_clusters}{Maximum number of clusters}

\item{use_advanced_diversity}{Whether to use advanced FFS algorithm}

\item{verbose}{Whether to print progress}
}
\value{
Vector of representative indices
}
\description{
Internal function to select cluster representatives that are diverse
across the similarity space.
}
\keyword{internal}
