% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{extract_parrot_networks}
\alias{extract_parrot_networks}
\title{Extract Network Structures for PARROT}
\usage{
extract_parrot_networks(strata)
}
\arguments{
\item{strata}{List of data domains}
}
\value{
List of network structures with adjacency matrices
}
\description{
Extract Network Structures for PARROT
}
\keyword{internal}
