% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feature_block_indices}
\alias{feature_block_indices}
\title{Compute feature block indices for concatenated matrices}
\usage{
feature_block_indices(domains)
}
\arguments{
\item{domains}{List of domain objects with element `$x`}
}
\value{
Named list of integer sequences matching the column ranges per block
}
\description{
Compute feature block indices for concatenated matrices
}
\keyword{internal}
