% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ot.R
\name{find_knn_indices}
\alias{find_knn_indices}
\title{Compute K-nearest neighbours under supported metrics}
\usage{
find_knn_indices(query, data, k = 5, metric = "euclidean")
}
\description{
Compute K-nearest neighbours under supported metrics
}
\keyword{internal}
