% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_grasp.R
\name{fit_pair.grasp_aligner}
\alias{fit_pair.grasp_aligner}
\title{Fit GRASP on a pair of domains (adapter)}
\usage{
\method{fit_pair}{grasp_aligner}(
  algo,
  X_i,
  X_j,
  links = NULL,
  ncomp = 30,
  q_descriptors = 100,
  sigma = 0.73,
  lambda = 0.1,
  use_laplacian = TRUE,
  solver = c("linear", "hungarian", "auction"),
  ...
)
}
\description{
Fit GRASP on a pair of domains (adapter)
}
