% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{fitted.fpgw}
\alias{fitted.fpgw}
\title{Extract fitted values}
\usage{
\method{fitted}{fpgw}(object, ...)
}
\arguments{
\item{object}{An fpgw object}

\item{...}{Additional arguments}
}
\value{
List of transport plans
}
\description{
Extract fitted values
}
