% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{gaussian_sim}
\alias{gaussian_sim}
\title{Calculate Gaussian Similarity from Data or Squared Distances}
\usage{
gaussian_sim(Z, sigma)
}
\arguments{
\item{Z}{Matrix of data (n x m) OR precomputed squared distances (n x n)}

\item{sigma}{Scaling parameter (sigma_P)}
}
\value{
n x n matrix of similarities
}
\description{
Calculate Gaussian Similarity from Data or Squared Distances
}
\keyword{internal}
