% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{get_parrot_use_cpp}
\alias{get_parrot_use_cpp}
\title{Get whether to use Rcpp implementations in PARROT}
\usage{
get_parrot_use_cpp()
}
\value{
Logical, whether to use Rcpp
}
\description{
Get whether to use Rcpp implementations in PARROT
}
