% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{get_parrot_use_rcpp}
\alias{get_parrot_use_rcpp}
\title{Get PARROT optimization flag}
\usage{
get_parrot_use_rcpp()
}
\value{
Logical indicating whether to use C++ optimizations
}
\description{
Get PARROT optimization flag
}
\keyword{internal}
