% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_multi.R
\name{global_geo_aligner}
\alias{global_geo_aligner}
\title{Global geometry aligner (native multi-view)}
\usage{
global_geo_aligner()
}
\value{
An object of class `c("global_geo_aligner", "aligner")` representing
  a global geometry alignment algorithm descriptor.
}
\description{
Global geometry aligner (native multi-view)
}
\examples{
algo <- global_geo_aligner()
class(algo)
}
