% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_multi.R
\name{gpa_aligner}
\alias{gpa_aligner}
\title{Generalized Procrustes aligner (native multi-view)}
\usage{
gpa_aligner()
}
\value{
An object of class `c("gpa_aligner", "aligner")` representing a
  Generalized Procrustes alignment algorithm descriptor.
}
\description{
Generalized Procrustes aligner (native multi-view)
}
\examples{
algo <- gpa_aligner()
aligner_capabilities(algo)
}
