% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca_align.R
\name{gpca_align.hyperdesign}
\alias{gpca_align.hyperdesign}
\title{Generalized PCA Alignment for Hyperdesign Data}
\usage{
gpca_align.hyperdesign(
  data,
  y,
  preproc = center(),
  ncomp = 2,
  simfun = neighborweights::binary_label_matrix,
  u = 0.5,
  lambda = 0.1,
  row_metric_scale = 1,
  control = gpca_align_control()
)
}
\arguments{
\item{data}{A hyperdesign object containing multiple data domains}

\item{y}{Name of the label variable to use for alignment}

\item{preproc}{Preprocessing function to apply to the data (default `center()`)}

\item{ncomp}{Number of components to extract (default 2)}

\item{simfun}{Function to compute similarity between labels}

\item{u}{Trade-off parameter between within-domain and between-domain similarity (0-1)}

\item{lambda}{Ridge regularisation added to the metric matrix}

\item{row_metric_scale}{Scaling applied to the assembled row metric before GPCA}

\item{control}{A list created by [gpca_align_control()] that tunes sparsification,
balancing, normalisation, and memory safeguards}
}
\description{
Generalized PCA Alignment for Hyperdesign Data
}
