% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca_align.R
\name{gpca_align}
\alias{gpca_align}
\title{Generalized PCA Alignment}
\usage{
gpca_align(data, ...)
}
\arguments{
\item{data}{A `hyperdesign` object or other supported input.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generalized PCA Alignment
}
