% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{grad_Jp}
\alias{grad_Jp}
\title{Calculate Gradient of Orthogonality Term (Jp) w.r.t W}
\usage{
grad_Jp(W)
}
\arguments{
\item{W}{Current weight matrix (d x m)}
}
\value{
Gradient matrix (d x m)
}
\description{
Calculate Gradient of Orthogonality Term (Jp) w.r.t W
}
\keyword{internal}
