% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{grad_Js}
\alias{grad_Js}
\title{Calculate Gradient of Similarity Term (Js) w.r.t W (Correct Implementation)}
\usage{
grad_Js(X, W, T, M, sigma, sum_M_batch)
}
\arguments{
\item{X}{Preprocessed data matrix (n x d) or batch (n_b x d)}

\item{W}{Current weight matrix (d x m)}

\item{T}{Target similarity matrix (n x n) or batch (n_b x n_b)}

\item{M}{Mask matrix (n x n) or batch (n_b x n_b)}

\item{sigma}{Scaling parameter (sigma_P)}

\item{sum_M_batch}{The sum of the *batch* mask matrix M (for consistent normalization)}
}
\value{
Gradient matrix (d x m)
}
\description{
Calculate Gradient of Similarity Term (Js) w.r.t W (Correct Implementation)
}
\keyword{internal}
