% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp_multiset.R
\name{gradient_single_multiset}
\alias{gradient_single_multiset}
\title{Gradient for Single Rotation in Multi-Graph Setting}
\usage{
gradient_single_multiset(basis_s, A_list, Ms, s, lambda, Gss = NULL)
}
\arguments{
\item{basis_s}{Spectral basis for graph s}

\item{A_list}{List of descriptor-basis products for all graphs}

\item{Ms}{List of all rotation matrices}

\item{s}{Index of current graph}

\item{lambda}{Regularization parameter}

\item{Gss}{Optional list of A_s^T A_s matrices for reuse}
}
\value{
Gradient matrix
}
\description{
Computes the gradient of the objective with respect to one rotation matrix
in the context of multiple graphs.
}
\keyword{internal}
