% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kema_scalable.R
\name{gram_Ls}
\alias{gram_Ls}
\title{Build Gram for same-class Laplacian using label factors.}
\usage{
gram_Ls(Zop, F)
}
\arguments{
\item{Zop}{Operator for kernel blocks.}

\item{F}{Label factor list from `label_factors()`.}
}
\value{
Symmetric dense matrix of size r x r.
}
\description{
Z^T L_s Z = Z^T D_s Z - (Z^T C)(Z^T C)^T
}
\examples{
\donttest{
library(Matrix)
K1 <- Matrix(rnorm(20), 10, 2)
Zop <- make_Zop_from_Ks(list(K1))
# F <- label_factors(...)  # Requires label factors
# G <- gram_Ls(Zop, F)
}
}
