% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp_multiset.R
\name{grasp_multiset.list}
\alias{grasp_multiset.list}
\title{GRASP Multiset for List of Matrices}
\usage{
\method{grasp_multiset}{list}(data, ...)
}
\arguments{
\item{data}{A list containing three or more matrices (nodes x features)}

\item{...}{Additional arguments passed to hyperdesign method}
}
\description{
GRASP Multiset for List of Matrices
}
