% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_grasp.R
\name{latent_dim.grasp_pair_fit}
\alias{latent_dim.grasp_pair_fit}
\title{Latent dimension for GRASP fit}
\usage{
\method{latent_dim}{grasp_pair_fit}(fit, ...)
}
\description{
Latent dimension for GRASP fit
}
