% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_framework.R
\name{latent_dim}
\alias{latent_dim}
\title{Latent dimension selected by the method (if applicable)}
\usage{
latent_dim(fit, ...)
}
\description{
Latent dimension selected by the method (if applicable)
}
