% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kema_scalable.R
\name{make_Zop_from_Ks}
\alias{make_Zop_from_Ks}
\title{Build matrix-free operator for block-diagonal kernel matrix constructed from
pre-computed kernel blocks.}
\usage{
make_Zop_from_Ks(Ks)
}
\arguments{
\item{Ks}{List of `Matrix` objects (each n_i x r_i) representing kernel
evaluations for a block.}
}
\value{
List with dimensions and forward/transpose operators.
}
\description{
Build matrix-free operator for block-diagonal kernel matrix constructed from
pre-computed kernel blocks.
}
