% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldalign-package.R
\docType{package}
\name{manifoldalign-package}
\alias{manifoldalign}
\alias{manifoldalign-package}
\title{manifoldalign package}
\description{
Internal package documentation and dynamic library registration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/manifoldalign/}
  \item \url{https://github.com/bbuchsbaum/manifoldalign}
  \item Report bugs at \url{https://github.com/bbuchsbaum/manifoldalign/issues}
}

}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
