% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mma_align_multiple.R
\name{mma_align_multiple.list}
\alias{mma_align_multiple.list}
\title{Multiset Manifold Alignment for a List of Matrices}
\usage{
\method{mma_align_multiple}{list}(data, ...)
}
\arguments{
\item{data}{A list with 3+ matrices (nodes x features)}

\item{...}{Passed to hyperdesign method}
}
\description{
Multiset Manifold Alignment for a List of Matrices
}
