% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{mma_align_multiple}
\alias{mma_align_multiple}
\title{Multiset Manifold Alignment (MMA)}
\usage{
mma_align_multiple(data, ...)
}
\arguments{
\item{data}{Input object; hyperdesign or list of matrices (see methods)}

\item{...}{Passed to method}
}
\description{
Align 3+ domains via spectral embeddings and EM-based probabilistic registration
with optional consensus template. See mma_align_multiple.hyperdesign for details.
}
