% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_grasp.R
\name{pair_loss.grasp_pair_fit}
\alias{pair_loss.grasp_pair_fit}
\title{Pairwise loss for GRASP fit}
\usage{
\method{pair_loss}{grasp_pair_fit}(fit, X_i = NULL, X_j = NULL, ...)
}
\description{
Pairwise loss for GRASP fit
}
