% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_parrot.R
\name{pair_loss.parrot_pair_fit}
\alias{pair_loss.parrot_pair_fit}
\title{Pairwise loss for PARROT fit}
\usage{
\method{pair_loss}{parrot_pair_fit}(fit, X_i = NULL, X_j = NULL, ...)
}
\description{
Pairwise loss for PARROT fit
}
