% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{pairwise_sqdist}
\alias{pairwise_sqdist}
\title{Calculate Pairwise Squared Euclidean Distances Efficiently}
\usage{
pairwise_sqdist(Z)
}
\arguments{
\item{Z}{Matrix (n x m)}
}
\value{
n x n matrix of squared distances
}
\description{
Calculate Pairwise Squared Euclidean Distances Efficiently
}
\keyword{internal}
