% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot-benchmark.R
\name{parrot-benchmark}
\alias{parrot-benchmark}
\title{PARROT Performance Benchmarking}
\value{
NULL (documentation page only).
}
\description{
Functions for benchmarking PARROT performance on large-scale networks
}
\keyword{internal}
