% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw.R
\name{partial_ot_mass}
\alias{partial_ot_mass}
\title{Mass-constrained partial OT oracle}
\usage{
partial_ot_mass(cost, p, q, rho, method = c("lpSolve", "Rsymphony"))
}
\value{
A numeric matrix representing the optimal transport plan.
}
\description{
Mass-constrained partial OT oracle
}
\keyword{internal}
