% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw.R
\name{partial_ot_tv}
\alias{partial_ot_tv}
\title{TV-penalized oracle}
\usage{
partial_ot_tv(cost, p, q, lambda)
}
\value{
A numeric matrix representing the optimal transport plan.
}
\description{
TV-penalized oracle
}
\keyword{internal}
