% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kema_scalable.R
\name{pivoted_chol_kernel_block}
\alias{pivoted_chol_kernel_block}
\title{Simple pivoted Cholesky kernel rank selection.}
\usage{
pivoted_chol_kernel_block(
  X,
  kernel,
  tol = 1e-06,
  max_rank = Inf,
  diag_fun = function(X) rep(1, nrow(X))
)
}
\arguments{
\item{X}{Data matrix (samples x features).}

\item{kernel}{kernlab kernel object.}

\item{tol}{Residual tolerance.}

\item{max_rank}{Optional rank cap.}

\item{diag_fun}{Optional shortcut for diagonal entries of kernel matrix.}
}
\value{
List with selected indices and factor matrix W (n x r).
}
\description{
Simple pivoted Cholesky kernel rank selection.
}
