% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{plot.fpgw}
\alias{plot.fpgw}
\title{Plot method for FPGW}
\usage{
\method{plot}{fpgw}(x, which = 1, pair = 1, ...)
}
\arguments{
\item{x}{An fpgw object}

\item{which}{Which plot to create (1: distance matrix, 2: transport plan)}

\item{pair}{For transport plan, which pair to plot (default: 1)}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
NULL, invisibly. Called for its side effect of producing a plot.
}
\description{
Plot method for FPGW
}
\examples{
\donttest{
set.seed(1)
X1 <- matrix(rnorm(30), 10, 3)
X2 <- matrix(rnorm(30), 10, 3)
library(multidesign)
md1 <- multidesign(X1, data.frame(id = 1:10))
md2 <- multidesign(X2, data.frame(id = 1:10))
hd <- hyperdesign(list(d1 = md1, d2 = md2))
res <- fpgw(hd)
plot(res, which = 1)
plot(res, which = 2, pair = 1)
}

}
