% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{predict.simembed}
\alias{predict.simembed}
\title{Predict Method for Similarity Embedding}
\usage{
\method{predict}{simembed}(object, newdata, ...)
}
\arguments{
\item{object}{A simembed object from linear_sim_embed()}

\item{newdata}{New data matrix (n_new x d) to project}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Matrix of projected coordinates (n_new x ncomp)
}
\description{
Projects new data using trained similarity embedding model.
}
\examples{
\donttest{
X <- matrix(rnorm(100 * 5), 100, 5)
model <- linear_sim_embed(X, ncomp = 2)
X_new <- matrix(rnorm(20 * 5), 20, 5)
Y_new <- predict(model, X_new)
}

}
