% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_utils.R, R/utils-ot.R
\name{prepare_ot_data}
\alias{prepare_ot_data}
\title{Prepare OT data from hyperdesign}
\usage{
prepare_ot_data(data, preproc = NULL)

prepare_ot_data(data, preproc = NULL)
}
\description{
Prepare OT data from hyperdesign

Extract data from hyperdesign and apply preprocessing
}
\keyword{internal}
