% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw.R
\name{print.fpgw}
\alias{print.fpgw}
\title{Print Method for FPGW Objects}
\usage{
\method{print}{fpgw}(x, ...)
}
\arguments{
\item{x}{An fpgw object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print Method for FPGW Objects
}
