% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{print.simembed}
\alias{print.simembed}
\title{Print Method for Similarity Embedding}
\usage{
\method{print}{simembed}(x, ...)
}
\arguments{
\item{x}{A simembed object}

\item{...}{Additional arguments}
}
\description{
Print Method for Similarity Embedding
}
