% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_utils.R
\name{project_onto_simplex_doubly_stochastic}
\alias{project_onto_simplex_doubly_stochastic}
\title{FPGW utility functions}
\usage{
project_onto_simplex_doubly_stochastic(P, p, q, max_iter = 100, tol = 1e-09)
}
\description{
FPGW utility functions
}
\keyword{Project}
\keyword{doubly}
\keyword{internal}
\keyword{matrices}
\keyword{onto}
\keyword{stochastic}
