% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_parrot.R
\name{relative_transform.parrot_pair_fit}
\alias{relative_transform.parrot_pair_fit}
\title{Extract relative transform from a PARROT pairwise fit}
\usage{
\method{relative_transform}{parrot_pair_fit}(fit, from = c("i", "j"), to = c("j", "i"), ...)
}
\description{
Extract relative transform from a PARROT pairwise fit
}
