% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{residuals.fpgw}
\alias{residuals.fpgw}
\title{Extract residuals (not applicable for FPGW)}
\usage{
\method{residuals}{fpgw}(object, ...)
}
\arguments{
\item{object}{An fpgw object}

\item{...}{Additional arguments}
}
\value{
NULL with a warning
}
\description{
Extract residuals (not applicable for FPGW)
}
\examples{
\donttest{
set.seed(1)
X1 <- matrix(rnorm(30), 10, 3)
X2 <- matrix(rnorm(30), 10, 3)
library(multidesign)
md1 <- multidesign(X1, data.frame(id = 1:10))
md2 <- multidesign(X2, data.frame(id = 1:10))
hd <- hyperdesign(list(d1 = md1, d2 = md2))
res <- fpgw(hd)
residuals(res)
}

}
