% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ot.R
\name{resolve_domain_index}
\alias{resolve_domain_index}
\title{Resolve domain input (index or name)}
\usage{
resolve_domain_index(x, domain_names)
}
\description{
Resolve domain input (index or name)
}
\keyword{internal}
