% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{set_parrot_use_rcpp}
\alias{set_parrot_use_rcpp}
\title{Set PARROT optimization flag}
\usage{
set_parrot_use_rcpp(use_rcpp = TRUE)
}
\arguments{
\item{use_rcpp}{Logical indicating whether to use C++ optimizations}
}
\description{
Set PARROT optimization flag
}
