% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw.R
\name{sinkhorn_fgw_warmstart}
\alias{sinkhorn_fgw_warmstart}
\title{Warm-start with entropic FGW}
\usage{
sinkhorn_fgw_warmstart(Cfeat, Cx, Cy, p, q, omega1, eps, iters)
}
\value{
A numeric matrix representing the warm-start transport plan.
}
\description{
Warm-start with entropic FGW
}
\keyword{internal}
