% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_align.R
\name{solve_assignment_cone}
\alias{solve_assignment_cone}
\title{Solve Assignment Problem for CONE-Align}
\usage{
solve_assignment_cone(Z1, Z2, Q, solver, anchors = NULL)
}
\arguments{
\item{Z1}{First embedding matrix}

\item{Z2}{Second embedding matrix}

\item{Q}{Orthogonal transformation matrix}

\item{solver}{Assignment algorithm ("linear" or "auction")}
}
\value{
Vector of assignment indices
}
\description{
Solve Assignment Problem for CONE-Align
}
\keyword{internal}
