% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp_multiset.R
\name{solve_diagonal_map}
\alias{solve_diagonal_map}
\title{Solve Diagonal Map to Anchor}
\usage{
solve_diagonal_map(desc, basis, M, Z)
}
\arguments{
\item{desc}{Descriptor matrix for graph}

\item{basis}{Spectral basis for graph}

\item{M}{Rotation matrix for graph}

\item{Z}{Anchor coefficient matrix}
}
\value{
Diagonal mapping matrix
}
\description{
Finds the diagonal matrix C that best maps from a graph's aligned
representation to the anchor representation.
}
\keyword{internal}
