% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_sylvester_rwr_cpp}
\alias{solve_sylvester_rwr_cpp}
\title{Solve Sylvester equation for RWR cost}
\usage{
solve_sylvester_rwr_cpp(W1, W2T, Cnode, beta, gamma, tol, max_iter)
}
\arguments{
\item{W1}{Transition matrix of network 1}

\item{W2T}{Transpose of transition matrix of network 2}

\item{beta}{RWR restart probability}

\item{gamma}{Cross-graph discount factor}

\item{tol}{Convergence tolerance}

\item{max_iter}{Maximum iterations}

\item{C_node}{Node-level cost matrix}
}
\value{
Solution matrix C_rwr
}
\description{
Iterative solution of: C_rwr = (1+β)C_node + (1-β)γ * W1 * C_rwr * W2^T
}
