% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_multi.R
\name{spectral_mnn_aligner}
\alias{spectral_mnn_aligner}
\title{Spectral MNN aligner (native multi-view)}
\usage{
spectral_mnn_aligner()
}
\value{
An object of class `c("spectral_mnn_aligner", "aligner")` representing
  a spectral MNN alignment algorithm descriptor.
}
\description{
Spectral MNN aligner (native multi-view)
}
\examples{
algo <- spectral_mnn_aligner()
aligner_capabilities(algo)
}
