% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_multi.R
\name{ssma_aligner}
\alias{ssma_aligner}
\title{SSMA aligner (native multi-view)}
\usage{
ssma_aligner()
}
\value{
An object of class `c("ssma_aligner", "aligner")` representing
  a semi-supervised manifold alignment (SSMA) algorithm descriptor.
}
\description{
SSMA aligner (native multi-view)
}
\examples{
algo <- ssma_aligner()
aligner_capabilities(algo)
}
