% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{summary.fpgw}
\alias{summary.fpgw}
\title{Summary method for FPGW}
\usage{
\method{summary}{fpgw}(object, ...)
}
\arguments{
\item{object}{An fpgw object}

\item{...}{Additional arguments}
}
\value{
The fpgw object, invisibly.
}
\description{
Summary method for FPGW
}
\examples{
\donttest{
set.seed(1)
X1 <- matrix(rnorm(30), 10, 3)
X2 <- matrix(rnorm(30), 10, 3)
library(multidesign)
md1 <- multidesign(X1, data.frame(id = 1:10))
md2 <- multidesign(X2, data.frame(id = 1:10))
hd <- hyperdesign(list(d1 = md1, d2 = md2))
res <- fpgw(hd)
summary(res)
}

}
