% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{transform.fpgw}
\alias{transform.fpgw}
\title{Transform new data using FPGW alignment}
\usage{
\method{transform}{fpgw}(x, newdata, source_index = 1, target_index = 2, ...)
}
\arguments{
\item{x}{An fpgw object}

\item{newdata}{New data to transform}

\item{source_index}{Index of source domain}

\item{target_index}{Index of target domain}

\item{...}{Additional arguments}
}
\value{
Transformed samples in the target domain feature space.
}
\description{
Transform new data using FPGW alignment
}
