% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kema_scalable.R
\name{Zt_times}
\alias{Zt_times}
\title{Apply Z^T to arbitrary multi-vectors without forming Z.}
\usage{
Zt_times(Zop, S)
}
\arguments{
\item{Zop}{Operator for Z.}

\item{S}{Dense or sparse matrix with n rows.}
}
\value{
Dense matrix with r rows.
}
\description{
Apply Z^T to arbitrary multi-vectors without forming Z.
}
