% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{adam_update}
\alias{adam_update}
\title{Adam optimizer (minimal implementation)}
\usage{
adam_update(
  par,
  grad,
  m_state,
  v_state,
  t,
  lr = 0.001,
  b1 = 0.9,
  b2 = 0.999,
  eps = 1e-08
)
}
\description{
Adam optimizer (minimal implementation)
}
\keyword{internal}
