% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolabel.R
\name{as.data.frame.pseudolabels}
\alias{as.data.frame.pseudolabels}
\title{Convert pseudolabels to data.frame}
\usage{
\method{as.data.frame}{pseudolabels}(x, ...)
}
\arguments{
\item{x}{A pseudolabels object}

\item{...}{Additional arguments (ignored)}
}
\description{
Convert pseudolabels to data.frame
}
