% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kema.R
\name{block_indices}
\alias{block_indices}
\title{Block indices for data list}
\usage{
block_indices(data_list)
}
\arguments{
\item{data_list}{List of data blocks}
}
\value{
Matrix with start and end indices for each block
}
\description{
Block indices for data list
}
