% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw.R
\name{classical_ot_lp_r}
\alias{classical_ot_lp_r}
\title{Classical OT linear program (R implementation)}
\usage{
classical_ot_lp_r(cost, p, q)
}
\value{
A numeric matrix representing the optimal transport plan.
}
\description{
Classical OT linear program (R implementation)
}
\keyword{internal}
