% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{coef.fpgw}
\alias{coef.fpgw}
\title{Extract coefficients from FPGW}
\usage{
\method{coef}{fpgw}(object, type = c("transport", "parameters"), ...)
}
\arguments{
\item{object}{An fpgw object}

\item{type}{Type of coefficients to extract}

\item{...}{Additional arguments}
}
\value{
Coefficients (transport plans or parameters)
}
\description{
Extract coefficients from FPGW
}
\examples{
\donttest{
set.seed(1)
X1 <- matrix(rnorm(30), 10, 3)
X2 <- matrix(rnorm(30), 10, 3)
library(multidesign)
md1 <- multidesign(X1, data.frame(id = 1:10))
md2 <- multidesign(X2, data.frame(id = 1:10))
hd <- hyperdesign(list(d1 = md1, d2 = md2))
res <- fpgw(hd)
transport_plans <- coef(res, type = "transport")
params <- coef(res, type = "parameters")
}

}
