% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_utils.R, R/utils-ot.R
\name{compute_feature_cost}
\alias{compute_feature_cost}
\title{Compute feature cost between domains}
\usage{
compute_feature_cost(X1, X2, metric = "euclidean")

compute_feature_cost(X1, X2, metric = "euclidean")
}
\description{
Compute feature cost between domains

Compute feature cost matrix between two domains
}
\keyword{internal}
