% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot_proximal.R
\name{compute_parrot_objective}
\alias{compute_parrot_objective}
\title{PARROT Proximal Point Method Implementation}
\usage{
compute_parrot_objective(
  S,
  C_rwr,
  networks,
  anchor_info,
  lambda_e,
  lambda_n,
  lambda_p
)
}
\arguments{
\item{S}{Transport plan matrix}

\item{C_rwr}{Position-aware cost matrix}

\item{networks}{List of network structures}

\item{anchor_info}{Anchor information}

\item{lambda_e}{Edge consistency weight}

\item{lambda_n}{Neighborhood consistency weight}

\item{lambda_p}{Anchor prior weight}
}
\value{
Objective function value
}
\description{
Correct implementation following Algorithm 2 from the paper
with proper objective function and monotonicity checks
}
\keyword{+}
\keyword{<C_rwr,}
\keyword{Compute}
\keyword{Computes}
\keyword{Function}
\keyword{Objective}
\keyword{PARROT}
\keyword{S>}
\keyword{full}
\keyword{internal}
\keyword{objective:}
\keyword{the}
\keyword{λ_e*L_e(S)}
\keyword{λ_n*L_n(S)}
\keyword{λ_p*L_a(S)}
