% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sim_embed.R
\name{.optimize_W_enhanced}
\alias{.optimize_W_enhanced}
\title{Enhanced ADAM optimizer with alpha scheduling}
\usage{
.optimize_W_enhanced(
  X,
  T,
  M,
  W0,
  sigma = 1,
  alpha_p = 0.2,
  alpha_schedule = FALSE,
  lr = 0.005,
  batch = 1,
  maxit = 500,
  tol = 1e-06,
  verbose = FALSE
)
}
\description{
Enhanced ADAM optimizer with alpha scheduling
}
\keyword{internal}
