% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters_parrot.R
\name{fit_pair.parrot_aligner}
\alias{fit_pair.parrot_aligner}
\title{Fit PARROT on a pair of domains (adapter)}
\usage{
\method{fit_pair}{parrot_aligner}(
  algo,
  X_i,
  X_j,
  links = NULL,
  ncomp = NULL,
  sigma = 0.15,
  lambda = 0.1,
  lambda_e = NULL,
  lambda_n = NULL,
  lambda_p = NULL,
  tau = 0.05,
  alpha = 0.2,
  gamma = 0.1,
  solver = c("sinkhorn"),
  max_iter = 100,
  tol = 1e-06,
  use_cpp = FALSE,
  anchors_policy = c("warn", "off", "error"),
  unsup_strategy = c("mnn_bootstrap", "attribute_only", "self_training"),
  conf_threshold = 0.9,
  max_rounds = 1,
  seed_strength = 0.05,
  laplacian_reg = NULL,
  ...
)
}
\description{
Fit PARROT on a pair of domains (adapter)
}
