% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{fitted.fpgw}
\alias{fitted.fpgw}
\title{Extract fitted values}
\usage{
\method{fitted}{fpgw}(object, ...)
}
\arguments{
\item{object}{An fpgw object}

\item{...}{Additional arguments}
}
\value{
List of transport plans
}
\description{
Extract fitted values
}
\examples{
\donttest{
set.seed(1)
X1 <- matrix(rnorm(30), 10, 3)
X2 <- matrix(rnorm(30), 10, 3)
library(multidesign)
md1 <- multidesign(X1, data.frame(id = 1:10))
md2 <- multidesign(X2, data.frame(id = 1:10))
hd <- hyperdesign(list(d1 = md1, d2 = md2))
res <- fpgw(hd)
plans <- fitted(res)
}

}
