% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ot.R
\name{get_plan_index}
\alias{get_plan_index}
\title{Compute linear index for upper-triangular pair}
\usage{
get_plan_index(n_domains, from_idx, to_idx)
}
\description{
Compute linear index for upper-triangular pair
}
\keyword{internal}
