% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kema_scalable.R
\name{gram_Ld}
\alias{gram_Ld}
\title{Build Gram for between-class Laplacian using label factors.}
\usage{
gram_Ld(Zop, F)
}
\arguments{
\item{Zop}{Operator for kernel blocks.}

\item{F}{Label factor list.}
}
\value{
Symmetric dense matrix.
}
\description{
Z^T L_d Z = Z^T D_d Z + (Z^T C)(Z^T C)^T - (Z^T ell)(Z^T ell)^T
}
