% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldalign-package.R
\docType{package}
\name{manifoldalign-package}
\alias{manifoldalign}
\alias{manifoldalign-package}
\title{manifoldalign package}
\description{
Internal package documentation and dynamic library registration.
}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
