% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{plot.fpgw}
\alias{plot.fpgw}
\title{Plot method for FPGW}
\usage{
\method{plot}{fpgw}(x, which = 1, pair = 1, ...)
}
\arguments{
\item{x}{An fpgw object}

\item{which}{Which plot to create (1: distance matrix, 2: transport plan)}

\item{pair}{For transport plan, which pair to plot (default: 1)}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot method for FPGW
}
