% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gromov_wasserstein.R
\name{predict.gromov_wasserstein}
\alias{predict.gromov_wasserstein}
\title{Predict method for Gromov-Wasserstein}
\usage{
\method{predict}{gromov_wasserstein}(object, newdata, from, to, ...)
}
\arguments{
\item{object}{A gromov_wasserstein object}

\item{newdata}{New data from one domain to transport to another}

\item{from}{Source domain index or name}

\item{to}{Target domain index or name}

\item{...}{Additional arguments}
}
\value{
Transported samples in the target domain space
}
\description{
Transport new samples using learned GW alignment
}
