% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw.R
\name{print.fpgw}
\alias{print.fpgw}
\title{Print Method for FPGW Objects}
\usage{
\method{print}{fpgw}(x, ...)
}
\arguments{
\item{x}{An fpgw object}

\item{...}{Additional arguments (currently unused)}
}
\value{
The fpgw object x, invisibly.
}
\description{
Print Method for FPGW Objects
}
\examples{
\donttest{
set.seed(1)
X1 <- matrix(rnorm(30), 10, 3)
X2 <- matrix(rnorm(30), 10, 3)
library(multidesign)
md1 <- multidesign(X1, data.frame(id = 1:10))
md2 <- multidesign(X2, data.frame(id = 1:10))
hd <- hyperdesign(list(d1 = md1, d2 = md2))
res <- fpgw(hd)
print(res)
}
}
