% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolabel.R
\name{print.pseudolabels}
\alias{print.pseudolabels}
\title{Print method for pseudolabels objects}
\usage{
\method{print}{pseudolabels}(x, ...)
}
\arguments{
\item{x}{A pseudolabels object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for pseudolabels objects
}
