% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_framework.R
\name{relative_transform}
\alias{relative_transform}
\title{Extract a relative transform from a pairwise fit}
\usage{
relative_transform(fit, from = c("i", "j"), to = c("j", "i"), ...)
}
\description{
Extract a relative transform from a pairwise fit
}
