% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{residuals.fpgw}
\alias{residuals.fpgw}
\title{Extract residuals (not applicable for FPGW)}
\usage{
\method{residuals}{fpgw}(object, ...)
}
\arguments{
\item{object}{An fpgw object}

\item{...}{Additional arguments}
}
\value{
NULL with a warning
}
\description{
Extract residuals (not applicable for FPGW)
}
