% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_align.R
\name{solve_assignment_cone}
\alias{solve_assignment_cone}
\title{Solve Assignment Problem for CONE-Align}
\usage{
solve_assignment_cone(embed1, embed2, solver)
}
\arguments{
\item{embed1}{First embedding matrix}

\item{embed2}{Second embedding matrix}

\item{solver}{Assignment solver method}
}
\value{
Assignment vector (1-based indexing)
}
\description{
Solve Assignment Problem for CONE-Align
}
\keyword{internal}
