% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_align.R
\name{solve_procrustes_cone}
\alias{solve_procrustes_cone}
\title{Solve Procrustes Problem for CONE-Align}
\usage{
solve_procrustes_cone(Z1, Z2, P, lambda = 0)
}
\arguments{
\item{Z1}{Source embedding matrix}

\item{Z2}{Target embedding matrix}

\item{P}{Permutation matrix (or vector of indices)}

\item{lambda}{Regularization parameter for Procrustes problem}
}
\value{
Orthogonal transformation matrix Q
}
\description{
Solve Procrustes Problem for CONE-Align
}
\keyword{internal}
