% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_align.R
\name{solve_procrustes_cone}
\alias{solve_procrustes_cone}
\title{Solve Procrustes Problem for CONE-Align}
\usage{
solve_procrustes_cone(source, target, lambda)
}
\arguments{
\item{source}{Source embedding matrix}

\item{target}{Target embedding matrix}

\item{lambda}{Regularization parameter}
}
\value{
Orthogonal rotation matrix
}
\description{
Solve Procrustes Problem for CONE-Align
}
\keyword{internal}
