% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpgw_methods.R
\name{summary.fpgw}
\alias{summary.fpgw}
\title{Summary method for FPGW}
\usage{
\method{summary}{fpgw}(object, ...)
}
\arguments{
\item{object}{An fpgw object}

\item{...}{Additional arguments}
}
\value{
Summary information
}
\description{
Summary method for FPGW
}
