% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudolabel.R
\name{summary.pseudolabels}
\alias{summary.pseudolabels}
\title{Summary method for pseudolabels objects}
\usage{
\method{summary}{pseudolabels}(object, ...)
}
\arguments{
\item{object}{A pseudolabels object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for pseudolabels objects
}
